requirement.txt
Click==7.0
cycler==0.10.0
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
Flask-Cors==3.0.6
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.1
monotonic==1.5
mysql-connector-python-rf==2.2.2
numpy==1.15.3
pandas==0.23.4
pyparsing==2.3.0
python-dateutil==2.7.5
python-engineio==2.3.2
python-socketio==2.0.0
pytz==2018.7
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
sklearn==0.0
Werkzeug==0.14.1
